try to fix build failure
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1866941 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/format/CellNumberFormatter.java
@@ -304,7 +304,7 @@ public class CellNumberFormatter extends CellFormatter {
     }
 
     private static int maxValue(List<Special> s) {
-        return (int) Math.round(Math.pow(10, s.size()) - 1);
+        return Math.toIntExact(Math.round(Math.pow(10, s.size()) - 1));
     }
 
     private List<Special> specialsFor(int pos, int takeFirst) {
